% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeting_octopus.R
\name{greeting_octopus}
\alias{greeting_octopus}
\title{Generate an octopus}
\usage{
greeting_octopus(include_angry = TRUE, color_funs = NULL)
}
\arguments{
\item{include_angry}{Should the angry expression be included as an option?}

\item{color_funs}{Optional, a list of \verb{crayon::*} functions to use
for the octopus' coloring (see examples).}
}
\description{
Generate an octopus
}
\examples{
# Disable the angry expression w/ `include_angry = FALSE`:
greeting_octopus(include_angry = FALSE)

# Override the default color options w/ `color_funs`:
greeting_octopus(color_funs = list(crayon::yellow, crayon::magenta))
greeting_octopus(color_funs = list(crayon::white))

}
